requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
google-auth==2.23.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
python-multipart==0.0.9
serverless-wsgi==3.0.5